+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Manual Step Reporting feature
 ### Fixed
 - A test status switch from PASSED to FAILED in case of an `@After` method failed
+- A test status when AssumptionViolatedException was thrown in `@Before` method
 
 ## [5.0.0-BETA-17]
 ### Fixed